% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get BER Data}
\usage{
get_data(time_series_code, output_format = c("codes", "names", "nested")[3])
}
\arguments{
\item{time_series_code}{time series code to return, \code{KBP7096B}}

\item{output}{options are ("codes", "names", "nested")}
}
\value{
tibble
}
\description{
Get timeseries by specifying vector of codes.
}
